Update SMain.c
Added timer delay for button debouncing.  Also globalized the data variable I was using for every Shift_In() call, which now just edits the global data variable.  Lastly, I added that when timermode ends it now flashes off and on to indicate that the timer is over and awaits for a certain button to be pressed to exit.  I'm probably not going to do anything after this commit since I'm using 90% of the program memory.