Update SMain.c
Made minor fix to Changetimedigit.  When the first digit is a 2, it will automatically make changes to the second digit when configuring the time for the 24 hour clock.  Also changed modulo to 4 for the second digit.  Turns out I forgot that you wrap around once you reach 24:00.  Changed some comments and got rid of the PORTA=0 line.