Update SMain.c
Biggest point of this update was removing the whichnumber() method, and replacing it with an array with those same values.  Now instead of calling a function it simply takes the correct value from the array that holds the binary representation of the numbers on the 7-segment display.  Also I removed some useless lines of code that did nothing.  Also when changing the time, I made it so it resets the tmr0 register and resets timer0count to prevent it from changing the time too fast.  This update freed up 1% more of the programmable memory.  This is most likely the last update I'll give for this project, but with the extra space, you probably could add a little more.